+++ src/main.cpp
@@ -19,14 +19,6 @@
 
 #include "utility.h"
 
-#include <Adafruit_BNO055.h>
-#include <Adafruit_Sensor.h>
-#include <Wire.h>
-// Math functions used to interpret IMU values
-#include "matrix.h"
-#include "quaternion.h"
-#include "vector.h"
-
 /*! @file svea_arduino_src.ino*/
 
 /*
@@ -235,42 +227,8 @@ void EncoderReadingToMsg(const encoders::encoder_reading_t &reading, lli_encoder
     msg.left_time_delta = reading.left_time_delta;
 }
 
-#define BNO055_SAMPLERATE_DELAY_MS (100)
-Adafruit_BNO055 bno = Adafruit_BNO055(55); //, BNO055_ADDRESS, &Wire1);
 
-bool setupIMU() {
-    if (!bno.begin()) {
-        /* There was a problem detecting the BNO055 ... check your connections */
-        Serial.print("Ooops, no BNO055 detected ... Check your wiring or I2C ADDR!");
-        return false;
-    }
-    bno.setExtCrystalUse(false);
-}
 
-void IMUReadingToMsg(lli_imu_t &msg) {
-    msg.header.stamp = nh.now();
-    msg.orientation.x = bno.getQuat().x();
-    msg.orientation.y = bno.getQuat().y();
-    msg.orientation.z = bno.getQuat().z();
-    msg.orientation.w = bno.getQuat().w();
-
-    imu::Vector<3> gyro = bno.getVector(Adafruit_BNO055::VECTOR_GYROSCOPE);
-    msg.angular_velocity.x = gyro.x();
-    msg.angular_velocity.y = gyro.y();
-    msg.angular_velocity.z = gyro.z();
-
-    imu::Vector<3> accel = bno.getVector(Adafruit_BNO055::VECTOR_ACCELEROMETER);
-    msg.linear_acceleration.x = accel.x();
-    msg.linear_acceleration.y = accel.y();
-    msg.linear_acceleration.z = accel.z();
-
-    float fakeCovariance = 0;
-    for (int i = 0; i < 9; ++i) {
-        msg.orientation_covariance[i] = fakeCovariance;
-        msg.angular_velocity_covariance[i] = fakeCovariance;
-        msg.linear_acceleration_covariance[i] = fakeCovariance;
-    }
-}
 
 /*!
  * @brief Steering callibration functionality. Should be called in every loop update.
@@ -425,7 +383,7 @@ void loop() {
         encoder_pub.publish(&MSG_ENCODER);
     }
 
-    IMUReadingToMsg(MSG_IMU);
+    IMUReadingToMsg(nh.now(), MSG_IMU);
     imu_pub.publish(&MSG_IMU);
 
     // PCB LED Logic
